:sparkles: more changes for working with analyzer output
