+++ README.md
@@ -10,27 +10,7 @@ Static reports to visualize data generated by Konveyor LSP Analyzer.
 npm ci
 ```
 
-- The UI allows to visualize the data generated by Analyzer LSP CLI but it does not generate it. Therefore we need to first generate the data using Analyzer CLI. Follow the instructions to run analyzer-lsp [here](https://github.com/konveyor/analyzer-lsp).
-
-The previous command generated a report folder. Let's set the enviroment variable `WINDUP_REPORT_PATH` to point to the generated report folder. E.g.
-
-```
-export WINDUP_REPORT_PATH=/home/username/my_app.jar.report
-```
-
-- The data that the UI will consume will come from the report generated in the previous step. Let's serve the data using NodeJS
-
-```bash
-node server.js
-```
-
-- Start the UI in development mode:
-
-```bash
-npm run start
-```
-
-- Access to the UI using http://localhost:3000
+The UI allows to visualize the data generated by Analyzer LSP CLI but it does not generate it. [Kantra CLI](https://github.com/konveyor-ecosystem/kantra) helps generate the data needed for the UI. Follow the instructions in [README](https://github.com/konveyor-ecosystem/kantra#readme) to generate the data.
 
 ## Code of Conduct
