--- public/apps.yaml
@@ -1,3 +1,0 @@
-- id: "1000"
-  name: Customer Tomcat Legacy
-  outputPath: /output.yaml
