--- public/files/home/pranav/Projects/example-applications/example-1/src/main/resources/persistence.properties
@@ -1,6 +1,0 @@
-jdbc.driverClassName=oracle.jdbc.driver.OracleDriver
-jdbc.url=jdbc:oracle:thin:@169.60.225.216:1521/XEPDB1
-jdbc.user=customers
-jdbc.password=customers
-hibernate.hbm2ddl.auto=create-drop
-hibernate.dialect=org.hibernate.dialect.OracleDialect
