+++ public/index.html
@@ -25,7 +25,7 @@
 
     <script src="./theme.js"></script>
     <script src="./version.js"></script>
-    <script src="./settings.js"></script>
+    <script src="./output.js"></script>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
