+++ public/output.js
@@ -1,0 +1,2 @@
+// empty file
+// place the output file generated by kantra here
