--- public/settings.js
@@ -1,4 +1,0 @@
-window["WINDUP_SETTINGS"] = {
-  forceOnline: false,
-  onlineApiUrl: "/api",
-};
