--- public/settings.js.ejs
@@ -1,7 +1,0 @@
-const basePath = window.location.pathname;
-const reportBasePath = basePath.substring(0, basePath.indexOf("/index.html")) + "/api";
-
-window["WINDUP_SETTINGS"] = {
-  forceOnline: true,
-  onlineApiUrl: reportBasePath,
-};
