--- public/windup.js
@@ -1,1 +1,0 @@
-// This file should contain all JSONs data discovered
