--- public/windup.js.ejs
@@ -1,1 +1,0 @@
-// Empty file
