+++ src/output.json
@@ -1,0 +1,1 @@
+[]
