+++ src/pages/application-details/pages/dependencies/dependencies.tsx
@@ -3,7 +3,7 @@ import { useOutletContext } from "react-router-dom";
 
 import { PageSection } from "@patternfly/react-core";
 
-import { ApplicationDto } from "@app/api/output";
+import { ApplicationDto } from "@app/api/report";
 import { DependenciesTable } from "@app/shared/components";
 
 export const Dependencies: React.FC = () => {
