--- src/queries/mocks/files/test-files/file1.java
@@ -1,16 +1,0 @@
-package com.example.apps;
-
-import io.fabric8.kubernetes.api.model.apiextensions.v1beta1.CustomResourceDefinition;
-
-public class App 
-{
-
-    public static void main( String[] args )
-    {
-        CustomResourceDefinition crd = new CustomResourceDefinition();
-        System.out.println( crd );
-
-        GenericClass<String> element = new GenericClass<String>("Hello world!");
-        element.get();
-    }
-}
