--- src/queries/mocks/files/test-files/file2.java
@@ -1,14 +1,0 @@
-package com.example.apps;
-
-
-public class GenericClass<T> {
-    private T element;
-
-    public GenericClass(T element) {
-        this.element = element;
-    }
-
-    public T get() {
-        return element;
-    }
-}
