+++ src/shared/components/violations-table/violations-table.tsx
@@ -31,10 +31,10 @@ import {
 } from "@patternfly/react-table";
 import { useDebounce } from "usehooks-ts";
 
-import { compareByCategoryFn } from "@app/api/output";
+import { compareByCategoryFn } from "@app/api/report";
 import { ALL_APPLICATIONS_ID } from "@app/Constants";
 import { FileProcessed, ViolationProcessed } from "@app/models/api-enriched";
-import { useAllApplications } from "@app/queries/ruleset";
+import { useAllApplications } from "@app/queries/report";
 import {
   SimpleTableWithToolbar,
   SimpleSelect,
